Alter build flows to have 1 for branches and 1 for tags.
Previous set-up was not adequate. This was due to a misunderstanding
of how GitHub Actions work. The new system attemps to have a special
flow that runs only when a tag is pushed. The code idea behind the
approach it that a tag corresponds 1-to-1 to a release.
Pushes of branches do not attempt to do releases.