Differentiate between 1) project version, 2) deb package version, and 3) rpm package version
The main version is 1). Ideally, this would be the only version which would follow semver.
This is impossible since deb and rpm packages impose specific rules about their versions
which is not compatible.
For example, the semver pre-release version '1.4.0-SNAPSHOT' would be interpreted by deb package
rules as version '1.4.0' with 'SNAPSHOT' being interpreted as the debian_revision.

For simplicity, this change defines the deb and rpm versions as the 'main' component of the semVer
project version.